requirements.txt
Django==1.9.7
djangorestframework==3.5.1
markdown==2.6.7
